% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadmlvar.R
\docType{data}
\name{ALL_NETWORK}
\alias{ALL_NETWORK}
\title{Constant for retreiving all available variables}
\format{An object of class \code{numeric} of length 1.}
\usage{
ALL_NETWORK
}
\description{
Constant for retreiving all available variables
}
\keyword{datasets}
