% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadmlvar.R
\name{between_network}
\alias{between_network}
\title{Calculate parameters of the network on the between-subject level
\code{between_network} Function that analyzes a dyadic timeseries network and
calculates density variables for between-subjects connections. Assumes variables
were provided as a list for one subject and then an equivalent list for another.}
\usage{
between_network(mlvar_model)
}
\arguments{
\item{mlvar_model}{An mlVAR object, usually created by the mlVAR function of
package mlVAR. The target variables should have been of even length, with
the first half belonging to part A of the dyad (e.g. men, clients,
children) and the other to part B (e.g. women, therapists, parents)}
}
\value{
a list of sample-level between-subjects density variables:
    intra.density - average density of both partners' between-subject networks
    intra_A.density - density of partner A's between-subject network
    intra_B.density - density of partner B's between-subject network
    inter.density - density of between-subject connections between partners
    ratio.density - ratio between inter density and average intra density
}
\description{
Calculate parameters of the network on the between-subject level
\code{between_network} Function that analyzes a dyadic timeseries network and
calculates density variables for between-subjects connections. Assumes variables
were provided as a list for one subject and then an equivalent list for another.
}
\examples{
\donttest{
require("mlVAR")
# creating a variable list with the first half containing variables for
# partner A and the second half containing variables for partner B
 var.names =
c("M_Anx.","M_Sad.","M_Vig.","M_Con.","W_Anx.","W_Sad.","W_Vig.","W_Con.")

# running mlVAR to create a model
fit1 <- mlVAR(sample1, vars = var.names, idvar = "ID", beepvar="DIARYDAY",
lags = 1,scale=TRUE,scaleWithin = TRUE)
}
between <- between_network(fit1)
}
