% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{CONTEMP}
\alias{CONTEMP}
\title{Constant for retreiving variables from the contemporaneous network}
\format{An object of class \code{numeric} of length 1.}
\usage{
CONTEMP
}
\description{
Constant for retreiving variables from the contemporaneous network
}
\keyword{datasets}
