% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{DENSITY}
\alias{DENSITY}
\title{Constant for retreiving densities of intra-partner networks, inter-partner
networks, and their ratio}
\format{An object of class \code{numeric} of length 1.}
\usage{
DENSITY
}
\description{
Constant for retreiving densities of intra-partner networks, inter-partner
networks, and their ratio
}
\keyword{datasets}
