% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadmlvar.R
\docType{package}
\name{dyadmlvar-package}
\alias{dyadmlvar}
\alias{dyadmlvar-package}
\title{dyadmlvar: Dyadic Timeseries Extension for Mlvar}
\description{
Analyze dyadic mlVAR networks, computing variables such as intra-partner and inter-partner node connections, strength of nodes, and density of the intra-partner and inter-partner networks.
}
\details{
The package analyzes dyadic networks computed by the mlVAR package.
  Use read_network to analyze a network, get_names to get lists of variable
  names for specific types of variables (e.g. inter-partner variables), and
  subset them from the table of all variables (ntwrk$all if ntwrk is the
  output of read_network). Due to the large number of variables you will
  probably want to use the lasso function to see which variables explain
  significant amounts of variance in some outcome vector. More detail can be
  found in Bar-Kalifa & Sened, 2019 (Under review).
}
\author{
\strong{Maintainer}: Haran Sened \email{haranse@gmail.com}

Authors:
\itemize{
  \item Eran Bar-Kalifa \email{eran.barkalifa@gmail.com}
}

}
