% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadmlvar.R
\docType{data}
\name{fit1}
\alias{fit1}
\title{Example mlVAR model}
\format{An mlVAR object with 6 nodes}
\source{
An internal dataset, analyzed using the mlVAR package (Epskamp,
  Deserno & Bringmann, 2018).
}
\usage{
fit1
}
\description{
An mlVAR model obtained from analyzing sample1, with some
  internal variables removed to conserve package space
}
\keyword{datasets}
