% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{INTER}
\alias{INTER}
\title{Constant for retreiving node connections between partner A variables to
parter B variables}
\format{An object of class \code{numeric} of length 1.}
\usage{
INTER
}
\description{
Constant for retreiving node connections between partner A variables to
parter B variables
}
\keyword{datasets}
