% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadmlvar.R
\docType{data}
\name{INTRA_A}
\alias{INTRA_A}
\title{Constant for retreiving node connections between partner A variables}
\format{An object of class \code{numeric} of length 1.}
\usage{
INTRA_A
}
\description{
Constant for retreiving node connections between partner A variables
}
\keyword{datasets}
