% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadmlvar.R
\docType{data}
\name{INTRA_B}
\alias{INTRA_B}
\title{Constant for retreiving node connections between partner B variables}
\format{An object of class \code{numeric} of length 1.}
\usage{
INTRA_B
}
\description{
Constant for retreiving node connections between partner B variables
}
\keyword{datasets}
