% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lasso}
\alias{lasso}
\title{Use a LASSO algorithm to find important predictors}
\usage{
lasso(Data, Predictors, Outcome, Seeds = 1, Train = F,
  PropOfTrain = 0.75)
}
\description{
\code{read_network} Function that uses the LASSO shrinkage reduction method
to find meaningful predictors of an outcome vector
}
