% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read_network}
\alias{read_network}
\title{Analyze a Dyadic Timeseries Network.}
\usage{
read_network(mlvar_model)
}
\description{
\code{read_network} Function that analyzes a dyadic timeseries network and calculates strength
and density variables. Assumes variables are list for one subject and then
equivalent list for another.
}
