% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadmlvar.R
\docType{data}
\name{sample1}
\alias{sample1}
\title{Example data}
\format{an mlVAR object with 6 nodes}
\source{
an internal dataset. For details on the data collection see:
  Bar-Kalifa, E., Rafaeli, E., & Sened, H. (2016). Truth and bias in daily
  judgments of support receipt between romantic partners. Personal
  Relationships, 23(1), 42-61.
}
\usage{
sample1
}
\description{
Diary data collected from 80 couples. Variables include daily
  feelings of anger, sadness and anxiety for each partner, an id for each
  couples and a diaryday denoting the relevant day of the diary (e.g. 0 for
  the first day, 34 for day 35 of the diary)
}
\keyword{datasets}
