% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadmlvar.R
\docType{data}
\name{sat}
\alias{sat}
\title{Relationship satisfaction data}
\format{a dataframe with 7 variables and 80 rows}
\source{
an internal dataset, for details on the data collection see:
  Bar-Kalifa, E., Rafaeli, E., & Sened, H. (2016). Truth and bias in daily
  judgments of support receipt between romantic partners. Personal
  Relationships, 23(1), 42-61.
}
\usage{
sat
}
\description{
relationship satisfaction before and after completing a daily
  diary for 80 couples, and residuals of post-diary satisfaction after
  adjusting for pre-diary satisfaction.
}
\keyword{datasets}
