% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{STRENGTH}
\alias{STRENGTH}
\title{Constant for retreiving total node strengths}
\format{An object of class \code{numeric} of length 1.}
\usage{
STRENGTH
}
\description{
Constant for retreiving total node strengths
}
\keyword{datasets}
