% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{TEMPORAL}
\alias{TEMPORAL}
\title{Constant for retreiving variables from the temporal network}
\format{An object of class \code{numeric} of length 1.}
\usage{
TEMPORAL
}
\description{
Constant for retreiving variables from the temporal network
}
\keyword{datasets}
