% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadmlvar.R
\name{read_network}
\alias{read_network}
\title{Analyze a Dyadic Timeseries Network.}
\usage{
read_network(mlvar_model)
}
\arguments{
\item{mlvar_model}{An mlVAR object, usually created by the mlVAR function of
package mlVAR. The target variables should have been of even length, with
the first half belonging to part A of the dyad (e.g. men, clients,
children) and the other to part B (e.g. women, therapists, parents)}
}
\value{
a dyadNetwork object with variables characterizing the network
}
\description{
\code{read_network} Function that analyzes a dyadic timeseries network and
calculates strength and density variables. Assumes variables are list for one
subject and then an equivalent list for another.
}
\examples{
\donttest{
require("mlVAR")
# creating a variable list with the first half containing variables for
# partner A and the second half containing variables for partner B
 var.names =
c("M_Anx.","M_Sad.","M_Vig.","M_Con.","W_Anx.","W_Sad.","W_Vig.","W_Con.")

# running mlVAR to create a model
fit1 <- mlVAR(sample1, vars = var.names, idvar = "ID", beepvar="DIARYDAY",
lags = 1,scale=TRUE,scaleWithin = TRUE)
}
ntwrk <- read_network(fit1)
}
